add CLI option to skip migration failures
To support this, each configuration migration had to be wrapped in a function call to
`exec_migration_func` that decides whether it will allow an exception to be thrown
or just log it so in effect skipping the failure. To test this feature, an HTTPadapter is
added that replaces HTTP status code to a failed one randomly.